% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{add_day_type}
\alias{add_day_type}
\title{Append a column with a given date's day type (weekday or weekend),
day of week, and day category (holiday, weekend, or weekday)}
\usage{
add_day_type(sensor_data)
}
\arguments{
\item{sensor_data}{data frame for single sensor returned from \code{pull_sensor()}}
}
\value{
The original data.table with additional columns
\itemize{
\item \code{day_type} either "Weekday" or "Weekend"
\item \code{day_of_week} one of "Monday", "Tuesday", etc.
\item \code{day_category} one of "Weekday", "Weekend", or "Holiday"
}
}
\description{
Append a column with a given date's day type (weekday or weekend),
day of week, and day category (holiday, weekend, or weekday)
}
