% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weather.R
\name{add_weather}
\alias{add_weather}
\title{Append weather data from Automated Surface Observing System (ASOS) via
the Iowa State University API.}
\usage{
add_weather(
  sensor_data,
  save_raw = FALSE,
  save_location = ".",
  interval_length = 1,
  station = "MSP",
  time_zone = "America\%2FChicago"
)
}
\arguments{
\item{sensor_data}{data.table, a data.table of aggregated sensor data.}

\item{save_raw}{logical, whether to save the raw data. Default is \code{FALSE}}

\item{save_location}{character, relative path to save location}

\item{interval_length}{nunmeric, interval length in hours. Default is \code{1}.}

\item{station}{character, ASOS station code. Default is \code{"MSP"} for the Minneapolis/St. Paul airport}

\item{time_zone}{character, time zone code. Default is \code{"America\%2FChicago"}.}
}
\value{
the original data.table with columns added:
\itemize{
\item \code{tmpf.mean} numeric, the mean temperature in Fahrenheit
\item \code{p01i.sum} numeric, the total precipitation in inches
\item \code{p01i.mean} numeric, the mean precipitation in inches
\item \code{sknt.mean} numeric, the mean wind speed in knots
\item \code{vsby.mean} numeric, the mean visibility distance in miles
}

See \href{http://mesonet.agron.iastate.edu/request/download.phtml?network=MN_ASOS}{Mesonet} for additional variables and descriptions.
}
\description{
Access temperature, precipitation, wind speed, and visibility
measures for a given ASOS station in Minnesota.
}
\details{
For additional Minnesota station ID codes, see
the \href{http://mesonet.agron.iastate.edu/sites/networks.php?network=MN_ASOS}{Mesonet station directory}.
}
\section{Deprecation}{

\strong{This function is deprecated and will be removed in version 0.3.0.}

The \code{add_weather()} function is no longer maintained. Users should
implement their own weather data workflows using alternative data sources
and packages.
}

\examples{
\dontrun{

library(tc.sensors)
library(dplyr)
config <- pull_configuration()

config_sample <- dplyr::filter(config, config$detector_abandoned == "f") \%>\%
  dplyr::sample_n(1)
yesterday <- as.Date(Sys.Date() - 365)

sensor_results <- pull_sensor(
  sensor = config_sample$detector_name[[1]],
  pull_date = yesterday
)

aggregate_sensor(sensor_results,
  interval_length = 1,
  config = config_sample
) \%>\%
  add_weather()
}
}
