% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_configuration.R
\name{attr_clean}
\alias{attr_clean}
\title{Clean node and detector attributes}
\usage{
attr_clean(category, attribute, metro_config)
}
\arguments{
\item{category}{string, one of "detector", "r_node", or "corridor"}

\item{attribute}{string. Options vary for each category.
Detector attributes
- "name"
- "label"
- "category"
- "lane"
- "field"
- "abandoned"
Node attributes
- "name"
- "n_type"
- "transition"
- "label"
- "lon"
- "lat"
- "lanes"
- "shift"
- "s_limit"
- "station_id"
- "attach_side"
Corridor attributes
- "route"
- "dir"}

\item{metro_config}{sensor configuration}
}
\value{
formatted configuration
}
\description{
Clean node and detector attributes
}
\keyword{internal}
