% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_speed}
\alias{calculate_speed}
\title{Calculate speed in miles per hour at different temporal aggregation levels}
\usage{
calculate_speed(sensor_data, configuration, interval_length = 1)
}
\arguments{
\item{sensor_data}{a data table of sensor data at 30 second intervals}

\item{intverval_length}{numeric, the interval length in hours. Default is \code{1}.
\code{0.25} indicates 15 minutes.}
}
\value{

}
\description{
Calculate speed in miles per hour at different temporal aggregation levels
}
\details{
There are 60 scans per second, which means there are 60*60 = 1,800 scans per
30-second interval. The occupancy value in the 30-second interval data
represents the number of scans that were occupied of the 1,800 scans in that
interval.

With 60 scans per second, 60 seconds per minute there are 3,600 scans per minute.
With 3,600 scans per minute, 60 minutes per hour there are 216,000 scans per hour.
To find the number of scans in 15 minutes, we can multiply 0.25 * 216000 = 54,000 scans.
}
\concept{calculations}
