% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{flag_impossible}
\alias{flag_impossible}
\title{Flag impossible values at given aggregation level}
\usage{
flag_impossible(sensor_data)
}
\arguments{
\item{sensor_data}{data frame returned from \code{pull_sensor()}}
}
\value{

}
\description{
Flag impossible values at given aggregation level
}
\section{Flag criteria}{
\preformatted{- Hourly
  - total hourly occupancy exceeds 216,000 scans
  - total hourly volume exceeds 2,300
- 30-sec
  - total 30-second volume exceeds 20 cars
  - total 30-second occupancy exceed 1,800
}
}

