% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayfly_utils.R
\name{mayfly_get_corridors}
\alias{mayfly_get_corridors}
\title{Get list of available corridors for a district}
\usage{
mayfly_get_corridors(district = "metro", year = NULL, date = NULL)
}
\arguments{
\item{district}{character, district code. Default is "metro"}

\item{year}{character or numeric, year to query. Optional.}

\item{date}{character, date in YYYYMMDD format. Optional.}
}
\value{
character vector of corridor names
}
\description{
Get list of available corridors for a district
}
\keyword{internal}
