% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayfly_utils.R
\name{mayfly_get_dates}
\alias{mayfly_get_dates}
\title{Get list of available dates for a district and year}
\usage{
mayfly_get_dates(district = "metro", year)
}
\arguments{
\item{district}{character, district code. Default is "metro"}

\item{year}{character or numeric, year to query}
}
\value{
character vector of available dates in YYYYMMDD format
}
\description{
Get list of available dates for a district and year
}
\keyword{internal}
