% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayfly_utils.R
\name{mayfly_get_districts}
\alias{mayfly_get_districts}
\title{Get list of available districts}
\usage{
mayfly_get_districts()
}
\value{
character vector of district codes
}
\description{
Get list of available districts
}
\keyword{internal}
