% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayfly_utils.R
\name{mayfly_get_years}
\alias{mayfly_get_years}
\title{Get list of available years for a district}
\usage{
mayfly_get_years(district = "metro")
}
\arguments{
\item{district}{character, district code. Default is "metro"}
}
\value{
character vector of available years
}
\description{
Get list of available years for a district
}
\keyword{internal}
