% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_sensor_range.R
\name{pull_sensor_range}
\alias{pull_sensor_range}
\title{Pull data from multiple sensors for a given date range}
\usage{
pull_sensor_range(start_date, end_date, config, use_furrr = TRUE, ...)
}
\arguments{
\item{start_date}{character, the start date in "YYYY-MM-DD" format}

\item{end_date}{character, the end date in "YYYY-MM-DD" format}

\item{config}{data.table, sensor configuration for multiple sensors}

\item{use_furrr}{logical, whether to use \code{{furrr}} for parallel processing}

\item{...}{Additional arguments to pass to \code{pull_sensor()}. Default is \code{TRUE}}
}
\value{
list, one data.table for each unique sensor.
}
\description{
Pull data from multiple sensors for a given date range
}
\examples{
\dontrun{
library(tc.sensors)

sensor_config <- pull_configuration()

pull_sensor_range(start_date = "2018-01-01",
end_date = "2018-01-01",
config = sensor_config)

}


}
