% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{replace_impossible}
\alias{replace_impossible}
\title{Replace impossible volume and occupancy values with \code{NA} at given interval}
\usage{
replace_impossible(sensor_data, interval_length = NA)
}
\arguments{
\item{sensor_data}{data frame for single sensor returned from \code{pull_sensor()}}

\item{interval_length}{numeric, the interval length in hours.
\code{NA} indicates no aggregation (30 second data)
\code{0.25} indicates 15 minutes.
Default is \code{1}.}
}
\value{
the original data.table with impossible volume and occupancy values
replaced with \code{NA}.
}
\description{
Replace impossible volume and occupancy values with \code{NA} at given interval
}
\details{
\subsection{Criteria}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{- 30-sec
  - Total 30-second volume exceeds 20 cars
  - Total 30-second occupancy exceed 1,800 scans
- Hourly
  - Total hourly occupancy exceeds 216,000 scans (60 scans per second \\* 60 secs per min \\*
    60 mins per hour = 216,000 scans per hour)
  - Total hourly volume exceeds 2,300 cars
- Percent nulls >= 10\%
}\if{html}{\out{</div>}}
}
}
