% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{scrub_sensor}
\alias{scrub_sensor}
\title{Scrub sensor level data}
\usage{
scrub_sensor(sensor_data, level)
}
\arguments{
\item{sensor_data}{data frame returned from \code{pull_sensor()}}

\item{level}{string, options are "raw", "fifteen", "hour"}
}
\value{

}
\description{
Scrub sensor level data
}
