% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sensor_data.R
\name{aggregate_sensor_data}
\alias{aggregate_sensor_data}
\title{Aggregate raw sensor data to a chosen level}
\usage{
aggregate_sensor_data(
  sensor_data,
  config,
  interval_length,
  calculations = c("volume.sum", "occupancy.sum")
)
}
\arguments{
\item{sensor_data}{data frame for single sensor returned from \code{pull_sensor()}}

\item{intverval_length}{numeric, the interval length in hours. Default is \code{1}.
\code{0.25} indicates 15 minutes.}
}
\value{

}
\description{
Aggregate raw sensor data to a chosen level
}
