% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_sensor.R
\name{extension_pull}
\alias{extension_pull}
\title{Pull extension}
\usage{
extension_pull(ext, ext_name, sensor, pull_date, quiet = TRUE)
}
\arguments{
\item{ext}{string, either \code{"v"} for volume or \code{"c"} for occupancy}

\item{ext_name}{string, either \code{"volume"} or \code{"occupancy"}}

\item{sensor}{character, the sensor ID.
See documentation for \code{\link{pull_sensor_ids}} to obtain metro sensor IDs.}

\item{pull_date}{character, the date of data to pull.
Needs to by in "YYYY-MM-DD" format.}

\item{quiet}{boolean, whether to hide messages. Default is TRUE}
}
\value{
a tibble
}
\description{
Pull extension
}
\keyword{internal}
