% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_sensor.R
\name{extension_pull}
\alias{extension_pull}
\title{Pull extension}
\usage{
extension_pull(ext, sensor, pull_date)
}
\arguments{
\item{ext}{either \code{"v"} for volume or \code{"c"} for occupancy}

\item{sensor}{an object of class integer or string which indicates the sensor ID.
See documentation for \code{\link{pull_sensor_ids}} to obtain metro sensor IDs.}

\item{pull_date}{an object of class string which indicates the date of data to pull.
Needs to by in "YYYY-MM-DD" format.}
}
\value{
a tibble
}
\description{
Pull extension
}
\keyword{internal}
