% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spatial_lines.R
\name{generate_spatial_lines}
\alias{generate_spatial_lines}
\title{Generate spatial lines dataset from sensor configuration}
\usage{
generate_spatial_lines(config, class = "sf")
}
\arguments{
\item{config}{data.table, sensor configuration for multiple sensors}

\item{class}{character, the output class. Default is \code{"sf"}.
Other option is \code{"sp"}.}
}
\value{
an \code{sp} or \code{sf} object with poly lines for each corridor
}
\description{
Generate spatial lines dataset from sensor configuration
}
