% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_configuration.R
\name{pull_sensor_ids}
\alias{pull_sensor_ids}
\title{Pull all sensor IDs in the Twin Cities metro}
\usage{
pull_sensor_ids(.quiet = TRUE)
}
\arguments{
\item{.quiet}{logical, whether to hide messages. Default is \code{TRUE}}
}
\value{
data.table containing variable "\code{detector}"
}
\description{
Create a data.table containing sensor IDs for MnDOT metro district, mainly to be used with \code{pull_sensor()}
}
\examples{
\dontrun{
sensors <- sensor_pull()
}
}
\seealso{
Other loop sensor functions: 
\code{\link{pull_configuration}()},
\code{\link{pull_sensor}()}
}
\concept{loop sensor functions}
