% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_sensor_ids.R
\name{pull_sensor_ids}
\alias{pull_sensor_ids}
\title{Function to pull all sensor IDs in the Twin Cities metro

Create a tidy dataframe containing sensor IDs for MnDOT metro district, mainly to be used with pull_sensor}
\usage{
pull_sensor_ids(.quiet = TRUE)
}
\arguments{
\item{.quiet}{logical, whether to hide messages. Default is TRUE}
}
\value{
dataframe containing variable "detector"
}
\description{
Function to pull all sensor IDs in the Twin Cities metro

Create a tidy dataframe containing sensor IDs for MnDOT metro district, mainly to be used with pull_sensor
}
\examples{
\dontrun{
sensors <- sensor_pull()
}
}
\seealso{
Other loop sensor functions: 
\code{\link{pull_configuration}()},
\code{\link{pull_sensor}()}
}
\concept{loop sensor functions}
