% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_sensor_functions.R
\name{pull_sensor_ids}
\alias{pull_sensor_ids}
\title{Function to pull all sensor IDs in the Twin Cities metro}
\usage{
pull_sensor_ids()
}
\value{
dataframe containing variable "detector"
}
\description{
Create a tidy dataframe containing sensor IDs for MnDOT metro district, mainly to be used with pull_sensor
}
\examples{
sensors <- sensor_pull()

}
