% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{replace_impossible}
\alias{replace_impossible}
\title{Replace impossible volume and occupancy values with \code{NA} at given interval}
\usage{
replace_impossible(sensor_data, interval_length = NA)
}
\arguments{
\item{sensor_data}{data frame for single sensor returned from \code{pull_sensor()}}

\item{interval_length}{numeric, the interval length in hours.
\code{NA} indicates no aggregation (30 second data)
\code{0.25} indicates 15 minutes.
Default is \code{1}.}
}
\value{
the original data.table with impossible volume and occupancy values
replaced with \code{NA}.
}
\description{
Replace impossible volume and occupancy values with \code{NA} at given interval
}
\section{Flag criteria}{
\preformatted{- Hourly
  - total hourly occupancy exceeds 216,000 scans
  - total hourly volume exceeds 2,300
- 30-sec
  - total 30-second volume exceeds 20 cars
  - total 30-second occupancy exceed 1,800
}

To find impossible values at any given interval length, we can multiply the
}

\author{
@R c(person("Ashley", "Asmus"),
person("Liz", "Roten"))
}
