% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{scrub_sensor}
\alias{scrub_sensor}
\title{Scrub sensor level data}
\usage{
scrub_sensor(sensor_data, interval_length)
}
\arguments{
\item{sensor_data}{data frame for single sensor returned from \code{pull_sensor()}}

\item{interval_length}{numeric, the interval length in hours.
\code{NA} indicates no aggregation (30 second data)
\code{0.25} indicates 15 minutes.
Default is \code{1}.}
}
\value{

}
\description{
Scrub sensor level data
}
