% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc.sensors-package.R
\docType{package}
\name{tc.sensors-package}
\alias{tc.sensors}
\alias{tc.sensors-package}
\title{tc.sensors: Retrieve Loop Detector Data from the MnDOT JSON Feed}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Process data collected from Minnesota Department of Transportation (MnDOT) loop detectors installed on the Minnesota Freeway system in 30-second interval measurements of occupancy and volume, data which are pushed daily to a public JSON feed. Occupancy and volume data can be used to calculate speed and delay.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Metropolitan-Council/tc.sensors/issues}
}

}
\author{
\strong{Maintainer}: Liz Roten \email{liz.roten@metc.state.mn.us} (\href{https://orcid.org/0000-0002-5346-3549}{ORCID})

Authors:
\itemize{
  \item Nicole Sullivan \email{nicole.sullivan@metc.state.mn.us}
}

Other contributors:
\itemize{
  \item Metropolitan Council [copyright holder]
  \item Ashley Asmus \email{ashley.asmus@metc.state.mn.us} (\href{https://orcid.org/0000-0001-5505-1372}{ORCID}) [contributor]
}

}
\keyword{internal}
